Change default faceted view to 'faceted-preview-items'
Added 'faceted-preview-items' View as the default view for Faceted
navigable objects instead of relaying on 'folder_listing' to be forever
compatible with eea.facetednavigation. This is also a more flexible solution
as you can customize each item display by providing a custom
'faceted-preview-item' z3 component for each of this items.